pyproject.toml
[tool.poetry]
name = "destiny-timelost"
version = "0.1.0"
description = "Python application to simplify map generation for Destiny 2 timelost event"
authors = ["Dmitry Figol <git@dmfigol.me>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
uvloop = "^0.14.0"
httpx = "^0.11.0"
aiofiles = "^0.4.0"

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
bpython = "^0.18"
black = "^19.10b0"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
